Add files via upload
Business Case #1: A document listing the equipment for a small publishing house with multiple departments located in a downtown skyscraper. The employees have hybrid work schedules. The company is small, but has brought the mainframe into the building for greater control over intellectual property. The company expects to grow into new markets in the next five years. The company partners with other small businesses in the publishing industry to focus on growth for all of their concerns.